fix($name) fix backtoppc name
