+++ embacktop/backtoppc/assets/style/backtoppc.scss
@@ -3,7 +3,7 @@
   bottom: 24px;
   right: 24px;
   transition: opacity 300ms linear;
-  opacity:0;
+  opacity: 0;
   z-index: 1;
 
   &-fadein {
@@ -13,7 +13,7 @@
 
   &-item {
     background: #fff;
-    box-shadow: 0 4px 16px -2px rgba(0,0,0,0.10);
+    box-shadow: 0 4px 16px -2px rgba(0, 0, 0, 0.1);
     border-radius: 2px;
     width: 80px;
     height: 80px;
@@ -42,8 +42,8 @@
     }
 
     &-qrcode {
-      background: #FFFFFF;
-      box-shadow: 0 4px 16px -2px rgba(0,0,0,0.10);
+      background: #fff;
+      box-shadow: 0 4px 16px -2px rgba(0, 0, 0, 0.1);
       border-radius: 2px;
       padding-top: 20px;
       position: absolute;
@@ -52,7 +52,7 @@
       width: 160px;
       padding-bottom: 16px;
       transition: opacity 500ms linear;
-      opacity:0;
+      opacity: 0;
 
       &-fadein {
         opacity: 100;
@@ -66,9 +66,9 @@
       }
 
       &-text {
-        font-family: PingFangSC-Regular;
+        font-family: 'PingFangSC-Regular';
         font-size: 12px;
-        color: #999999;
+        color: #999;
         text-align: center;
         line-height: 12px;
       }
