+++ embacktop/embacktop.scss
@@ -1,1 +1,1 @@
-@import 'backtop/assets/style/backtop.scss';
+@import 'backtoppc/assets/style/backtoppc';
