+++ package.json
@@ -139,7 +139,6 @@
     ]
   },
   "pre-commit": [
-    "lint-staged",
-    "test"
+    "lint-staged"
   ]
 }
